Move frontend into folder, expr to ded. file
Expressions will start getting more complex here in a bit
in order to support predicates and ternary logic, so that
seems like a good time to make space for some tests.
